+++ packages/builder/src/pages/builder/workspace/[application]/agent/[agentId]/knowledge/sharepoint/DisplaySharePointSiteModal.svelte
@@ -1,18 +1,12 @@
 <script lang="ts">
-  import {
-    ActionButton,
-    Body,
-    Modal,
-    ModalContent,
-    TreeView,
-  } from "@budibase/bbui"
+  import { Body, Modal, ModalContent, TreeView } from "@budibase/bbui"
   import {
     AgentKnowledgeSourceType,
     KnowledgeBaseFileStatus,
     type KnowledgeBaseFile,
   } from "@budibase/types"
   import { agentsStore, selectedAgent } from "@/stores/portal"
-  import SharePointEntryTreeItem from "./SharePointEntryTreeItem.svelte"
+  import SharePointEntryTreeItem from "./tree/SharePointEntryTreeItem.svelte"
   import { buildEntryTree } from "./sharePointModalUtils"
 
   export interface Props {
@@ -23,7 +17,6 @@
 
   let { agentId, siteId, onEdit }: Props = $props()
 
-  let editing = $state(false)
   let modal = $state<Modal>()
 
   const sharePointSource = $derived.by(() => {
@@ -112,83 +105,45 @@
     if (!siteId || !onEdit) {
       return
     }
-    editing = true
-    try {
-      await onEdit(siteId)
-      hide()
-    } finally {
-      editing = false
-    }
+
+    await onEdit(siteId)
+    hide()
   }
 </script>
 
 <Modal bind:this={modal}>
   <ModalContent
-    custom
-    showCloseIcon={false}
+    title={`SharePoint - ${selectedSiteLabel}`}
     showDivider={false}
-    showConfirmButton
-    confirmText="Close"
-    cancelText="Close"
     disabled={!sourceId}
     onConfirm={hide}
     onCancel={hide}
+    size="XL"
+    showConfirmButton={false}
+    showCancelButton={false}
+    showSecondaryButton
+    secondaryButtonText="Edit files"
+    secondaryAction={handleEdit}
   >
-    <div class="content">
-      <div class="title">
-        <Body size="S">SharePoint - {selectedSiteLabel}</Body>
+    {#if entryTree.length === 0}
+      <Body size="S">
+        {showProcessingState
+          ? "SharePoint sync is in progress. Files will appear here shortly."
+          : "No selected files found for this site."}
+      </Body>
+    {:else}
+      <div class="entries-list">
+        <TreeView standalone={false} quiet width="auto" selectable={false}>
+          {#each entryTree as node (node.path)}
+            <SharePointEntryTreeItem {node} />
+          {/each}
+        </TreeView>
       </div>
-
-      <div class="entries-header">
-        <Body size="M">Synced files</Body>
-        <ActionButton quiet size="S" loading={editing} on:click={handleEdit}>
-          Edit files
-        </ActionButton>
-      </div>
-
-      {#if entryTree.length === 0}
-        <Body size="S">
-          {showProcessingState
-            ? "SharePoint sync is in progress. Files will appear here shortly."
-            : "No selected files found for this site."}
-        </Body>
-      {:else}
-        <div class="entries-list">
-          <TreeView width="100%" standalone={false} quiet>
-            {#each entryTree as node (node.path)}
-              <SharePointEntryTreeItem {node} />
-            {/each}
-          </TreeView>
-        </div>
-      {/if}
-    </div>
+    {/if}
   </ModalContent>
 </Modal>
 
 <style>
-  .content {
-    padding: var(--spacing-l);
-    width: min(760px, 95vw);
-    display: flex;
-    flex-direction: column;
-    gap: var(--spacing-xs);
-  }
-
-  .title {
-    padding-bottom: var(--spacing-xxs);
-  }
-
-  .entries-header {
-    margin-top: var(--spacing-xxs);
-    display: flex;
-    align-items: center;
-    gap: var(--spacing-xs);
-  }
-
-  .entries-header :global(.spectrum-ActionButton:first-of-type) {
-    margin-left: auto;
-  }
-
   .entries-list {
     margin-top: var(--spacing-xxs);
     max-height: 420px;
